survey: Add tabularization of question 25.: how many projects tracked
Adds subroutine normalize_number used to extract and tabularize number
from free-form expression. It uses first number or range; for range it
uses value in the middle of range. Both '<n>-<m>' and '<n> to <m>' are
supported for ranges. Return value is number or range.

Use this subroutine to tabularize free-form question:
  '25. How many different projects do you manage using GIT?'

NOTE: normalize_number is optimized for question 25.

Signed-off-by: Jakub Narebski <jnareb@gmail.com>