Move information about survey from code to YAML file
Instead of embedding information about survey (sections, questions and
their type, answers, etc.) in the code, put it in editable YAML file.
This way it should be easier to move from analyzing survey for 2009,
to analyzing survey for 2010 year.

Because code cannot be reliably and sanely embedded in what is to be a
data file, there was added new subroutine to add pre- and
post-processing subroutines after reading data from YAML file.

While at it, read sections from @survey_data, instead of embedding
them.

NOT TESTED (but it runs `--help` without warnings).

Signed-off-by: Jakub Narebski <jnareb@gmail.com>