Survs.com: Generate graphical histogram in 'wiki' output format
When using 'wiki' (Mediawiki) output format, add fourth column with
color bar of width representing percentage of responses; visibly it
looks like vertical graphical histogram of responses.

It can be turned off with `--no-graph`.  The customization of the look
of this graph is not yet exported, meaning that it is not available as
options/option arguments from command line.

Because we need base, the prototype / calling convention of
fmt_row_percent changed to include base as fourth parameter (this
means that we don't have to pass percentage as parameter, but it could
be calculated by fmt_row_percent itself - this is left for later
commit).

The 'matrix' questions fo not get graphical histogram (yet?).

Signed-off-by: Jakub Narebski <jnareb@gmail.com>